+++ code/packages/sdk/src/nodes/group/group.ts
@@ -10,6 +10,7 @@ import {
 } from '@inditextech/weave-types';
 import { WeaveNode } from '../node';
 import { WEAVE_GROUP_NODE_TYPE } from './constants';
+import type { WeaveNodesSelectionPlugin } from '@/plugins/nodes-selection/nodes-selection';
 
 export class WeaveGroupNode extends WeaveNode {
   protected nodeType: string = WEAVE_GROUP_NODE_TYPE;
@@ -32,6 +33,13 @@ export class WeaveGroupNode extends WeaveNode {
     nodeInstance.setAttrs({
       ...nextProps,
     });
+
+    const nodesSelectionPlugin =
+      this.instance.getPlugin<WeaveNodesSelectionPlugin>('nodesSelection');
+
+    if (nodesSelectionPlugin) {
+      nodesSelectionPlugin.getTransformer().forceUpdate();
+    }
   }
 
   serialize(instance: WeaveElementInstance): WeaveStateElement {