+++ code/packages/sdk/src/nodes/rectangle/rectangle.ts
@@ -9,6 +9,7 @@ import {
 } from '@inditextech/weave-types';
 import { WeaveNode } from '../node';
 import { WEAVE_RECTANGLE_NODE_TYPE } from './constants';
+import type { WeaveNodesSelectionPlugin } from '@/plugins/nodes-selection/nodes-selection';
 
 export class WeaveRectangleNode extends WeaveNode {
   protected nodeType: string = WEAVE_RECTANGLE_NODE_TYPE;
@@ -31,5 +32,12 @@ export class WeaveRectangleNode extends WeaveNode {
     nodeInstance.setAttrs({
       ...nextProps,
     });
+
+    const nodesSelectionPlugin =
+      this.instance.getPlugin<WeaveNodesSelectionPlugin>('nodesSelection');
+
+    if (nodesSelectionPlugin) {
+      nodesSelectionPlugin.getTransformer().forceUpdate();
+    }
   }
 }